convert request to join to tailwind
