+++ packages/client/modules/teamDashboard/containers/Team/TeamContainer.tsx
@@ -4,7 +4,7 @@ import {PreloadedQuery, usePreloadedQuery} from 'react-relay'
 import {Route} from 'react-router'
 import {matchPath, Switch} from 'react-router-dom'
 import ErrorBoundary from '~/components/ErrorBoundary'
-import RequestToJoinComponent from '../../../../components/RequestToJoinComponent'
+import RequestToJoinComponent from '../../../../components/RequestToJoin'
 import useRouter from '../../../../hooks/useRouter'
 import {TeamContainerQuery} from '../../../../__generated__/TeamContainerQuery.graphql'
 import Team from '../../components/Team/Team'
